backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
ping3==4.0.8
dnspython==2.7.0
speedtest-cli==2.1.3
pydantic==2.10.6